PWN-7616 - Move state machine to bridge package
