+++ app/src/main/java/org/p2p/wallet/bridge/send/statemachine/SendFeatureException.kt
@@ -1,4 +1,4 @@
-package org.p2p.wallet.newsend.statemachine
+package org.p2p.wallet.bridge.send.statemachine
 
 import java.math.BigDecimal
