+++ app/src/main/java/org/p2p/wallet/bridge/send/statemachine/SendState.kt
@@ -1,8 +1,8 @@
-package org.p2p.wallet.newsend.statemachine
+package org.p2p.wallet.bridge.send.statemachine
 
 import java.math.BigDecimal
-import org.p2p.wallet.newsend.statemachine.model.SendFee
-import org.p2p.wallet.newsend.statemachine.model.SendToken
+import org.p2p.wallet.bridge.send.statemachine.model.SendFee
+import org.p2p.wallet.bridge.send.statemachine.model.SendToken
 
 sealed interface SendBridgeState
