+++ app/src/main/java/org/p2p/wallet/bridge/send/statemachine/handler/bridge/AmountChangeActionHandler.kt
@@ -1,18 +1,18 @@
-package org.p2p.wallet.newsend.statemachine.handler.bridge
+package org.p2p.wallet.bridge.send.statemachine.handler.bridge
 
 import java.math.BigDecimal
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flatMapMerge
 import kotlinx.coroutines.flow.flow
 import org.p2p.core.utils.isZero
-import org.p2p.wallet.newsend.statemachine.SendActionHandler
-import org.p2p.wallet.newsend.statemachine.SendFeatureAction
-import org.p2p.wallet.newsend.statemachine.SendState
-import org.p2p.wallet.newsend.statemachine.bridgeToken
-import org.p2p.wallet.newsend.statemachine.fee
-import org.p2p.wallet.newsend.statemachine.fee.SendBridgeFeeLoader
-import org.p2p.wallet.newsend.statemachine.mapper.SendBridgeStaticStateMapper
-import org.p2p.wallet.newsend.statemachine.validator.SendBridgeValidator
+import org.p2p.wallet.bridge.send.statemachine.SendActionHandler
+import org.p2p.wallet.bridge.send.statemachine.SendFeatureAction
+import org.p2p.wallet.bridge.send.statemachine.SendState
+import org.p2p.wallet.bridge.send.statemachine.bridgeToken
+import org.p2p.wallet.bridge.send.statemachine.fee
+import org.p2p.wallet.bridge.send.statemachine.fee.SendBridgeFeeLoader
+import org.p2p.wallet.bridge.send.statemachine.mapper.SendBridgeStaticStateMapper
+import org.p2p.wallet.bridge.send.statemachine.validator.SendBridgeValidator
 
 class AmountChangeActionHandler(
     private val mapper: SendBridgeStaticStateMapper,
