+++ app/src/main/java/org/p2p/wallet/bridge/send/statemachine/handler/bridge/InitFeatureActionHandler.kt
@@ -1,13 +1,13 @@
-package org.p2p.wallet.newsend.statemachine.handler.bridge
+package org.p2p.wallet.bridge.send.statemachine.handler.bridge
 
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flatMapMerge
 import kotlinx.coroutines.flow.flow
-import org.p2p.wallet.newsend.statemachine.SendActionHandler
-import org.p2p.wallet.newsend.statemachine.SendFeatureAction
-import org.p2p.wallet.newsend.statemachine.SendState
-import org.p2p.wallet.newsend.statemachine.fee.SendBridgeFeeLoader
-import org.p2p.wallet.newsend.statemachine.model.SendInitialData
+import org.p2p.wallet.bridge.send.statemachine.SendActionHandler
+import org.p2p.wallet.bridge.send.statemachine.SendFeatureAction
+import org.p2p.wallet.bridge.send.statemachine.SendState
+import org.p2p.wallet.bridge.send.statemachine.fee.SendBridgeFeeLoader
+import org.p2p.wallet.bridge.send.statemachine.model.SendInitialData
 
 class InitFeatureActionHandler(
     private val feeLoader: SendBridgeFeeLoader,
