+++ app/src/main/java/org/p2p/wallet/bridge/send/statemachine/model/SendInitialData.kt
@@ -1,13 +1,13 @@
-package org.p2p.wallet.newsend.statemachine.model
+package org.p2p.wallet.bridge.send.statemachine.model
 
 import java.math.BigDecimal
-import org.p2p.core.token.SolAddress
+import org.p2p.core.wrapper.eth.EthAddress
 
 sealed interface SendInitialData {
 
     data class Bridge(
         val initialToken: SendToken.Bridge,
         val initialAmount: BigDecimal?,
-        val recipient: SolAddress,
+        val recipient: EthAddress,
     )
 }
